create a new separate class FileReader
